Changes
Condensed three scripts into one
