+++ Assets/Scripts/Richard Foldder/Dialogue.cs
@@ -22,3 +22,46 @@ public class Dialogue
   
 }
 
+public class DialogueNode
+{
+    public int nodeId;
+    public string text;
+    public List<OptionNode> options;
+
+    public DialogueNode()
+    {
+        options = new List<OptionNode>();
+    }
+
+    public void addDialogue(string text, int id)
+    {
+        this.text = text;
+        this.nodeId = id;
+    }
+
+    public void addOption(string text, int dest)
+    {
+        OptionNode op = null;
+
+        op = new OptionNode(text, dest);
+        this.options.Add(op);
+
+
+
+    }
+}
+
+public class OptionNode
+{
+    public string text;
+    public int destId;
+
+    public OptionNode() { }
+
+    public OptionNode(string text, int dest)
+    {
+        this.text = text;
+        this.destId = dest;
+    }
+}
+
