--- Assets/Scripts/Richard Foldder/DialogueNode.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3daf20e814cd26b4fb0c123a127682c8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
