--- Assets/Scripts/Richard Foldder/OptionNode.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8c7bdfdcfd93849458f089011b4220bc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
